requirements.txt
flask==3.0.2
gunicorn==21.2.0
google-genai==0.3.0
python-dotenv==1.0.1
flask-cors==4.0.0
pytest==8.0.2
requests==2.31.0
